feat: create nft basic setup
