+++ components/ui/Base/BasePage/BasePage.types.d.ts
@@ -1,3 +1,3 @@
 export interface IBasePageProps {
-  children: string | JSX.Element | JSX.Element[];
+  children: any;
 }
