+++ components/ui/CreateNftCard/CreateNftCard.types.d.ts
@@ -1,0 +1,4 @@
+export interface ICreateNftCard {
+  onSingleClick?: () => void;
+  onCollectionClick?: () => void;
+}
