+++ pages/create-nft/CreateCollectionFrom.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const CreateCollectionFrom = () => {
+  return (
+    <div>CreateCollectionFrom</div>
+  )
+}
+
+export default CreateCollectionFrom
